backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pytz
RPi.GPIO
uwsgi
